Move block value validation
