+++ divi/src/BlockIncentivesPopulator.cpp
@@ -80,3 +80,22 @@ void BlockIncentivesPopulator::FillBlockPayee(CMutableTransaction& txNew, const 
         masternodePayments_.FillBlockPayee(txNew, payments, fProofOfStake);
     }
 }
+
+bool IsBlockValueValid(const I_SuperblockHeightValidator& heightValidator, const CBlockRewards &nExpectedValue, CAmount nMinted, int nHeight)
+{
+    auto nExpectedMintCombined = nExpectedValue.nStakeReward + nExpectedValue.nMasternodeReward;
+
+    // here we expect treasury block payment
+    if(heightValidator.IsValidTreasuryBlockHeight(nHeight)) {
+        nExpectedMintCombined += (nExpectedValue.nTreasuryReward + nExpectedValue.nCharityReward);
+    }
+    else if(heightValidator.IsValidLotteryBlockHeight(nHeight)) {
+        nExpectedMintCombined += nExpectedValue.nLotteryReward;
+    }
+
+    if (nMinted > nExpectedMintCombined) {
+        return false;
+    }
+
+    return true;
+}
