+++ divi/src/SuperblockHelpers.cpp
@@ -22,22 +22,3 @@ const I_BlockSubsidyProvider& SuperblockSubsidyContainer::blockSubsidiesProvider
 {
     return *blockSubsidies_;
 }
-
-bool IsBlockValueValid(const I_SuperblockHeightValidator& heightValidator, const CBlockRewards &nExpectedValue, CAmount nMinted, int nHeight)
-{
-    auto nExpectedMintCombined = nExpectedValue.nStakeReward + nExpectedValue.nMasternodeReward;
-
-    // here we expect treasury block payment
-    if(heightValidator.IsValidTreasuryBlockHeight(nHeight)) {
-        nExpectedMintCombined += (nExpectedValue.nTreasuryReward + nExpectedValue.nCharityReward);
-    }
-    else if(heightValidator.IsValidLotteryBlockHeight(nHeight)) {
-        nExpectedMintCombined += nExpectedValue.nLotteryReward;
-    }
-
-    if (nMinted > nExpectedMintCombined) {
-        return false;
-    }
-
-    return true;
-}
